Renamed "atomic site" to "site"
